Select and show message in room selected
