+++ components/Messages/ChatRoomsItem.tsx
@@ -8,6 +8,7 @@ import {
 } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 import ChevronRightIcon from '@material-ui/icons/ChevronRight'
+import NextLink from 'next/link'
 import React from 'react'
 import { Room } from '../../interfaces/chat/room.interface'
 
@@ -18,34 +19,36 @@ type ChatRoomsProps = {
 export function ChatRoomsItem({ room }: ChatRoomsProps) {
   const classes = useStyles({ isActive: room._id })
   return (
-    <ListItem divider button>
-      <ListItemAvatar className={classes.avatarSmall}>
-        <Badge
-          overlap="circle"
-          color="secondary"
-          badgeContent=" "
-          classes={{
-            badge: classes.badge,
-          }}
-          anchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'right',
-          }}
-          variant="dot"
-        >
-          <Avatar
-            alt={`Room logo`}
-            src={
-              'https://dcassetcdn.com/design_img/1559024/551167/551167_7840631_1559024_911ff84c_image.png'
-            }
-          />
-        </Badge>
-      </ListItemAvatar>
-      <ListItemText primary={room.name} />
-      <ListItemSecondaryAction>
-        <ChevronRightIcon />
-      </ListItemSecondaryAction>
-    </ListItem>
+    <NextLink href={`/messages/${room._id}`} passHref>
+      <ListItem divider button>
+        <ListItemAvatar className={classes.avatarSmall}>
+          <Badge
+            overlap="circle"
+            color="secondary"
+            badgeContent=" "
+            classes={{
+              badge: classes.badge,
+            }}
+            anchorOrigin={{
+              vertical: 'bottom',
+              horizontal: 'right',
+            }}
+            variant="dot"
+          >
+            <Avatar
+              alt={`Room logo`}
+              src={
+                'https://dcassetcdn.com/design_img/1559024/551167/551167_7840631_1559024_911ff84c_image.png'
+              }
+            />
+          </Badge>
+        </ListItemAvatar>
+        <ListItemText primary={room.name} />
+        <ListItemSecondaryAction>
+          <ChevronRightIcon />
+        </ListItemSecondaryAction>
+      </ListItem>
+    </NextLink>
   )
 }
